removing an item from the basket, reducing the item in the basket

+++ pizza/src/components/CartItem.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { useDispatch } from "react-redux";
-import { addItem, removeItem } from "../redux/slices/cartSlice";
+import { addItem, minusItem, removeItem } from "../redux/slices/cartSlice";
 
 const CartItem = ({ id, title, type, price, count, imageUrl }) => {
     const dispatch = useDispatch();
@@ -10,8 +10,14 @@ const CartItem = ({ id, title, type, price, count, imageUrl }) => {
     };
 
     const onClickMinus = () => {
-        dispatch(removeItem(id));
+        dispatch(minusItem(id));
     };
+    const onClickRemove = () => {
+        if (window.confirm("Are you sure you want to remove")) {
+            dispatch(removeItem(id));
+        }
+    };
+
     return (
         <div className="cart__item">
             <div className="cart__item-img">
@@ -66,7 +72,7 @@ const CartItem = ({ id, title, type, price, count, imageUrl }) => {
                 <b>{price * count} ₽</b>
             </div>
             <div className="cart__item-remove">
-                <div className="button button--outline button--circle">
+                <div onClick={onClickRemove} className="button button--outline button--circle">
                     <svg
                         width="10"
                         height="10"
